old-tests/submission/Microsoft/video/video_016.htm

<!DOCTYPE HTML>
<html>
<head>
  <title>HTML5 Media Elements: 'duration' returns NaN when no media data is available.</title>
  <meta content="text/html; charset=UTF-8" http-equiv="Content-Type">
  <link rel="author" title="Microsoft" href="http://www.microsoft.com/" />
  <link rel="help" href="http://www.w3.org/TR/html5/video.html#dom-media-duration" />
  <meta name="assert" content="'duration' returns NaN when no media data is available." />
</head>
<body>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>

    <div id=log></div>
    <video id="video0">Your browser does not support media elements.</video>
  
    <script type="text/javascript">

        test(function() {assert_true(isNaN(document.getElementById("video0").duration))}, "'duration' returns NaN when no media data is available.");

  </script>
</div>
</body>
</html>
